Refactor: Split the CLI module into a package
The aislop scanner flags cli.py on six counts: the file itself at
4188 lines, four over-long functions (merge at 663, blocked at 245,
_handle_org_merge at 229, _handle_repo_merge at 223) and a seven
parameter confirmation helper.

Following c22d70b, cli.py becomes a package whose __init__ re-exports
the module's entire top-level surface, private helpers included, so
every existing import and monkeypatch target still resolves. This was
verified by comparing the parsed top-level names of the original
module against the new package: 149 names, none missing, none added.
No test file needed changing.

The split follows the command structure. Each console command keeps
its signature, defaults and docstring untouched and moves only its
body; the long functions become orchestrators over named helpers, as
in 4e73f67. The command modules are imported inside an isort-off
block so Typer's registration order, and therefore --help, survives
import sorting. The --help output of the application and of every
subcommand is byte-identical to main.

merge_concurrency moves onto _MergeContext's derived state, which
brings _handle_org_preview_confirmation within the parameter budget
and removes it from _execute_org_confirmed_merge too.

Two incidental duplications collapse: merge built the same
_MergeContext three times, and repeated the same exception ladder
three times, now _run_guarded.

python -m dependamerge.cli keeps working through a new __main__, and
the dependamerge console script still resolves to cli:app.

merge_manager.py is deliberately untouched and its own findings are
left for a separate change.
