+++ src/dependamerge/cli/__main__.py
@@ -1,0 +1,8 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: 2026 The Linux Foundation
+"""Entry point for ``python -m dependamerge.cli``."""
+
+from . import app
+
+if __name__ == "__main__":
+    app()
